Solucion de Errores, de gestion Fabricantes

+++ Obligatorio1/Obligatorio1/Persistencia/pFabricante.cs
@@ -69,14 +69,33 @@ namespace Obligatorio1.Persistencia
         }
         public bool Baja(int pId)
         {
-            return Conexion.Instancia.InicializarConsulta("delete from Fabricantes where id_Fabricante=" + pId);
+            if (!this.EstaFabricanteEnOtraTabla(pId)) // si no es clave foranea en otra tabla se da de baja
+            {
+                return Conexion.Instancia.InicializarConsulta("delete from Fabricantes where id_Fabricante=" + pId);
+              
+            }
+            return false;
         }
+
+
 
         public bool Modificar(Fabricante pFabricante)
         {
             return Conexion.Instancia.InicializarConsulta("Exec ModificarFabricante " + pFabricante.Id + ",'" + pFabricante.Nombre + "','" + pFabricante.Direccion +
                                                            "','" + pFabricante.CorreoElectronico + "'");
         }
+        private bool EstaFabricanteEnOtraTabla(int pId)
+        {
+            string consulta = "select f. * from Fabricantes f  where f.id_fabricante =" + pId + " and f.id_Fabricante in (select id_Fabricante from Articulos)";
+            
+            DataSet datos = Conexion.Instancia.InicializarSeleccion(consulta);
+            if (datos.Tables[0].Rows.Count > 0)
+            {
+                return true;
+            }
+            return false;
+        
+    }
 
         public List<Fabricante> Listar()
         {
